third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-mode-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-mode with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-mode">
<meta name="assert" content="mask-border-mode supports only the grammar 'luminance | alpha'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-border-mode", "auto");
test_invalid_value("mask-border-mode", "none");
</script>
</body>
</html>
